matrix: fetch replied to messages and quote them
this is a WIP, i guess. when a message is replied to on matrix, the bridge now
fetches its content and adds quote marks in front before passing the thing on.

unwanted side effect: while testing with a matrix-matrix bridge, this kept
adding levels of quotes. that will get incredibly noisy.

though since i'm doing this for the xmpp-matrix bridge of some friends,
it might work differently.